+++ client/src/ts/modules/account/accountDetails.ts
@@ -1,4 +1,5 @@
 import { handleAuthSessionDestroyed, handleAuthSessionExpired } from "../global/authUtils";
+import { ConfirmModal } from "../global/ConfirmModal";
 import { getFullDateSTring } from "../global/dateTimeUtils";
 import ErrorSpan from "../global/ErrorSpan";
 import { AsyncErrorData, getAsyncErrorData } from "../global/errorUtils";
@@ -6,9 +7,10 @@ import { InfoModal } from "../global/InfoModal";
 import LoadingModal from "../global/LoadingModal";
 import popup from "../global/popup";
 import revealPassword from "../global/revealPassword";
+import { removeSignInCookies } from "../global/signOut";
 import { validateCode, validateConfirmPassword, validateDisplayName, validateEmail, validateNewPassword, validatePassword } from "../global/validation";
-import { resendDeletionEmailService, resendEmailUpdateEmailService, startAccountDeletionService, startEmailUpdateService, updateDisplayNameService, updatePasswordService } from "../services/accountServices";
-import { displayRequestSuspendedInfoModal, handleAccountLocked, handleOngoingOpposingRequest, handleOngoingRequest, handleRequestSuspended } from "./accountUtils";
+import { confirmAccountDeletionService, confirmEmailUpdateService, resendDeletionEmailService, resendEmailUpdateEmailService, startAccountDeletionService, startEmailUpdateService, updateDisplayNameService, updatePasswordService } from "../services/accountServices";
+import { handleAccountLocked, handleOngoingOpposingRequest, handleOngoingRequest, handleRequestSuspended } from "./accountUtils";
 import { accountState } from "./initAccount";
 
 type DetailsUpdateFormPurpose = 'emailUpdate' | 'displayNameUpdate' | 'passwordUpdate' | 'deleteAccount';
@@ -17,17 +19,11 @@ type ConfirmationFormPurpose = 'confirmEmailUpdate' | 'confirmAccountDeletion';
 interface AccountDetailsState {
   detailsUpdateFormPurpose: DetailsUpdateFormPurpose | null,
   confirmationFormPurpose: ConfirmationFormPurpose | null,
-
-  emailUpdateSuspensionExpiryTimestamp: number | null,
-  accountDeletionSuspensionExpiryTimestamp: number | null,
 };
 
-export const accountDetailsState: AccountDetailsState = {
+const accountDetailsState: AccountDetailsState = {
   detailsUpdateFormPurpose: null,
   confirmationFormPurpose: null,
-
-  emailUpdateSuspensionExpiryTimestamp: null,
-  accountDeletionSuspensionExpiryTimestamp: null,
 };
 
 const detailsElement: HTMLDivElement | null = document.querySelector('#details');
@@ -146,7 +142,38 @@ async function handleDetailsUpdateFormSubmission(e: SubmitEvent): Promise<void> 
 async function handleConfirmationFormSubmission(e: SubmitEvent): Promise<void> {
   e.preventDefault();
 
-  // TODO: implement
+  if (!accountDetailsState.confirmationFormPurpose) {
+    return;
+  };
+
+  if (accountDetailsState.confirmationFormPurpose === 'confirmEmailUpdate') {
+    await confirmEmailUpdate();
+    return;
+  };
+
+  const confirmModal: HTMLDivElement = ConfirmModal.display({
+    title: 'Are you sure want to delete your account?',
+    description: `This action is irreversible. You won't be able to recover your data if you proceed.`,
+    confirmBtnTitle: 'Delete account',
+    cancelBtnTitle: 'Cancel',
+    extraBtnTitle: null,
+    isDangerousAction: true,
+  });
+
+  confirmModal.addEventListener('click', async (e: MouseEvent) => {
+    if (!(e.target instanceof HTMLButtonElement)) {
+      return;
+    };
+
+    if (e.target.id === 'confirm-modal-confirm-btn') {
+      await confirmAccountDeletion();
+      return;
+    };
+
+    if (e.target.id === 'confirm-modal-cancel-btn') {
+      ConfirmModal.remove();
+    };
+  });
 };
 
 async function updateDisplayName(): Promise<void> {
@@ -382,13 +409,6 @@ async function startEmailUpdate(): Promise<void> {
     return;
   };
 
-  if (accountDetailsState.emailUpdateSuspensionExpiryTimestamp) {
-    displayRequestSuspendedInfoModal('email update', accountDetailsState.emailUpdateSuspensionExpiryTimestamp);
-    LoadingModal.remove();
-
-    return;
-  };
-
   if (accountDetailsState.confirmationFormPurpose === 'confirmEmailUpdate') {
     popup(`There's already an ongoing email update request.`, 'error');
     LoadingModal.remove();
@@ -444,7 +464,7 @@ async function startEmailUpdate(): Promise<void> {
 
     InfoModal.display({
       title: 'Email update process started.',
-      description: `You'll receive an email within the next 30 seconds with a confirmation code.`,
+      description: `An email will be sent to your new email address within the next 30 seconds with a confirmation code.`,
       btnTitle: 'Okay',
     }, { simple: true });
 
@@ -578,18 +598,120 @@ async function resendEmailUpdateEmail(): Promise<void> {
   };
 };
 
-async function startAccountDeletion(): Promise<void> {
+async function confirmEmailUpdate(): Promise<void> {
   LoadingModal.display();
 
-  if (!accountState.data) {
+  if (!accountState.data || !confirmationCodeInput) {
     popup('Something went wrong.', 'error');
     LoadingModal.remove();
 
     return;
   };
 
-  if (accountDetailsState.accountDeletionSuspensionExpiryTimestamp) {
-    displayRequestSuspendedInfoModal('account deletion', accountDetailsState.accountDeletionSuspensionExpiryTimestamp);
+  const isValidConfirmationCode: boolean = validateCode(confirmationCodeInput);
+
+  if (!isValidConfirmationCode) {
+    popup('Invalid confirmation code.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  const confirmationCode: string = confirmationCodeInput.value.toUpperCase();
+
+  try {
+    const { newEmail, authSessionCreated } = (await confirmEmailUpdateService({ confirmationCode })).data;
+
+    accountState.data.accountDetails.email = newEmail;
+    accountState.data.accountDetails.ongoing_email_update_request = false;
+
+    accountDetailsState.confirmationFormPurpose = null;
+
+    renderAccountDetails();
+    renderConfirmationForm();
+
+    popup('Email updated.', 'success');
+    LoadingModal.remove();
+
+    if (authSessionCreated) {
+      return;
+    };
+
+    const infoModal: HTMLDivElement = InfoModal.display({
+      title: 'Email successfully updated.',
+      description: `You'll just have to sign back into your account.`,
+      btnTitle: 'Okay',
+    });
+
+    infoModal.addEventListener('click', (e: MouseEvent) => {
+      if (!(e.target instanceof HTMLButtonElement)) {
+        return;
+      };
+
+      if (e.target.id === 'info-modal-info-btn') {
+        window.location.href = 'sign-in';
+      };
+    });
+
+  } catch (err: unknown) {
+    console.log(err);
+    LoadingModal.remove();
+
+    const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
+
+    if (!asyncErrorData) {
+      popup('Something went wrong.', 'error');
+      return;
+    };
+
+    const { status, errMessage, errReason, errResData } = asyncErrorData;
+
+    popup(errMessage, 'error');
+
+    if (status === 404) {
+      accountState.data.accountDetails.ongoing_email_update_request = false;
+      accountDetailsState.confirmationFormPurpose = null;
+
+      renderConfirmationForm();
+      return;
+    };
+
+    if (status === 403) {
+      handleRequestSuspended(errResData, 'email update');
+      return;
+    };
+
+    if (status === 401) {
+      if (errReason === 'incorrectCode') {
+        ErrorSpan.display(confirmationCodeInput, errMessage);
+        return;
+      };
+
+      if (errReason === 'requestSuspended') {
+        handleRequestSuspended(errResData, 'email update');
+        return;
+      };
+
+      if (errReason === 'authSessionExpired') {
+        handleAuthSessionExpired();
+        return;
+      };
+
+      handleAuthSessionDestroyed();
+      return;
+    };
+
+    if (status === 400 && errReason === 'confirmationCode') {
+      ErrorSpan.display(confirmationCodeInput, errMessage);
+    };
+  };
+};
+
+async function startAccountDeletion(): Promise<void> {
+  LoadingModal.display();
+
+  if (!accountState.data) {
+    popup('Something went wrong.', 'error');
     LoadingModal.remove();
 
     return;
@@ -764,6 +886,88 @@ async function resendDeletionEmail(): Promise<void> {
       };
 
       handleAuthSessionDestroyed();
+    };
+  };
+};
+
+async function confirmAccountDeletion(): Promise<void> {
+  LoadingModal.display();
+
+  if (!accountState.data || !confirmationCodeInput) {
+    popup('Something went wrong.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  const isValidConfirmationCode: boolean = validateCode(confirmationCodeInput);
+
+  if (!isValidConfirmationCode) {
+    popup('Invalid confirmation code.', 'error');
+    LoadingModal.remove();
+
+    return;
+  };
+
+  const confirmationCode: string = confirmationCodeInput.value.toUpperCase();
+
+  try {
+    await confirmAccountDeletionService(confirmationCode);
+    removeSignInCookies();
+
+    popup('Account deleted.', 'error');
+    setTimeout(() => window.location.href = 'home', 2000);
+
+  } catch (err: unknown) {
+    console.log(err);
+    LoadingModal.remove();
+
+    const asyncErrorData: AsyncErrorData | null = getAsyncErrorData(err);
+
+    if (!asyncErrorData) {
+      popup('Something went wrong.', 'error');
+      return;
+    };
+
+    const { status, errMessage, errReason, errResData } = asyncErrorData;
+
+    popup(errMessage, 'error');
+
+    if (status === 404) {
+      accountState.data.accountDetails.ongoing_account_deletion_request = false;
+      accountDetailsState.confirmationFormPurpose = null;
+
+      renderConfirmationForm();
+      return;
+    };
+
+    if (status === 403) {
+      handleRequestSuspended(errResData, 'account deletion');
+      return;
+    };
+
+    if (status === 401) {
+      if (errReason === 'incorrectCode') {
+        ErrorSpan.display(confirmationCodeInput, errMessage);
+        return;
+      };
+
+      if (errReason === 'requestSuspended') {
+        handleRequestSuspended(errResData, 'account deletion');
+        return;
+      };
+
+      if (errReason === 'authSessionExpired') {
+        handleAuthSessionExpired();
+        return;
+      };
+
+      handleAuthSessionDestroyed();
+      return;
+    };
+
+    if (status === 400 && errReason === 'confirmationCode') {
+      ErrorSpan.display(confirmationCodeInput, errMessage);
     };
   };
 };