+++ client/src/ts/modules/account/accountUtils.ts
@@ -1,7 +1,6 @@
 import { getDateAndTimeString } from "../global/dateTimeUtils";
 import { InfoModal } from "../global/InfoModal";
 import { removeSignInCookies } from "../global/signOut";
-import { accountDetailsState } from "./accountDetails";
 
 export function removeLoadingSkeleton(): void {
   document.querySelector('#loading-skeleton')?.remove();
@@ -60,7 +59,7 @@ export function handleOngoingOpposingRequest(ongoingRequestTitle: string,): void
   }, { simple: true });
 };
 
-export function handleRequestSuspended(errResData: unknown, ongoingRequestTitle: string): void {
+export function handleRequestSuspended(errResData: unknown, ongoingRequestTitle: 'email update' | 'account deletion'): void {
   if (typeof errResData !== 'object' || errResData === null) {
     return;
   };
@@ -73,21 +72,9 @@ export function handleRequestSuspended(errResData: unknown, ongoingRequestTitle:
     return;
   };
 
-  if (ongoingRequestTitle === 'account deletion') {
-    accountDetailsState.accountDeletionSuspensionExpiryTimestamp = errResData.expiryTimestamp;
-  };
-
-  if (ongoingRequestTitle === 'email update') {
-    accountDetailsState.emailUpdateSuspensionExpiryTimestamp = errResData.expiryTimestamp;
-  };
-
-  displayRequestSuspendedInfoModal(ongoingRequestTitle, errResData.expiryTimestamp);
-};
-
-export function displayRequestSuspendedInfoModal(requestTitle: string, suspensionExpiryTimestamp: number): void {
   InfoModal.display({
     title: `Request suspended.`,
-    description: `Your ${requestTitle} request has been suspended due to too many failed attempts\nYou can start a new one after ${suspensionExpiryTimestamp}.`,
+    description: `Your ${ongoingRequestTitle} request has been suspended due to too many failed attempts\nYou can start a new one after ${errResData.expiryTimestamp}.`,
     btnTitle: 'Okay',
   }, { simple: true });
 };