+++ client/src/ts/modules/services/accountServices.ts
@@ -144,6 +144,7 @@ interface ConfirmEmailUpdateBody {
 };
 
 interface ConfirmEmailUpdateData {
+  newEmail: string,
   authSessionCreated: boolean,
 };
 
@@ -166,7 +167,7 @@ export function resendDeletionEmailService(): Promise<AxiosResponse> {
 // --- --- ---
 
 export function confirmAccountDeletionService(confirmationCode: string): Promise<AxiosResponse> {
-  return axios.delete(`${accountsApiUrl}/deletion/start?confirmationCode=${confirmationCode}`);
+  return axios.delete(`${accountsApiUrl}/deletion/confirm?confirmationCode=${confirmationCode}`);
 };
 
 // --- --- ---